requirements.txt
# Core
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
PyYAML>=6.0

# Testing
pytest>=7.3.0

# Optional: MCMC (built-in fallback provided if missing)
emcee>=3.1.0
